Add files via upload
Interacted with Falcon_7b. Attempted to implement langchain functionalities for demo day test (even though we won't use this model unless it is miraculously uploaded to the cloud). The purpose of this experiment was to try and build functional code structures which could be implemented with langchain for the sake of memory and diagram generation.